Improve safety, type consistency, and code clarity
Updated `PKImageEncode_SetXMPMetadata_WMP` in `JXRGlueJxr.c` to use `strcpy_s` for safer string copying and improved error handling.

Changed loop variables in `test_pixels.cpp` to `uint32_t` for type consistency with dimension getters.

Suppressed unused variable warnings in `test_subblockattachment.cpp` by explicitly casting `index` to `(void)` in lambda functions.

Ensured type safety in `test_writer.cpp` by using `static_cast<uint32_t>` for width and height comparisons.

+++ Src/libCZI_UnitTests/test_writer.cpp
@@ -1328,7 +1328,7 @@ TEST(CziWriter, Writer13)
     EXPECT_TRUE(sbBlkInfo.GetCompressionMode() == CompressionMode::JpgXr) << "Incorrect result";
 
     auto bitmap = sbBlk->CreateBitmap();
-    EXPECT_TRUE(bitmap->GetWidth() == widthCompressedBitmap && bitmap->GetHeight() == heightCompressedBitmap) << "Incorrect result";
+    EXPECT_TRUE(bitmap->GetWidth() == static_cast<uint32_t>(widthCompressedBitmap) && bitmap->GetHeight() == static_cast<uint32_t>(heightCompressedBitmap)) << "Incorrect result";
     EXPECT_TRUE(bitmap->GetPixelType() == PixelType::Bgr24) << "Incorrect result";
 }
